Schema: add unique indicies on job_id and filename
This is to prevent the same file being listed more than once
as an input or output for a given job.  In general for the JSA
multiple jobs may share an input or output file, even though
that isn't the case for the SCUBA-2 HEALPix processing we
plan first to use this database for.