Web: improve pagination of job list page
The pagination routines have been extracted into a function
calculate_pagination in the jsa_proc.web.util module and a macro
render_page_control in the "macros.html" file.  This allows
the page controls to be displayed at both ends of the job list
page, and potentially also used for other pages.  The
default number of items per page for the job list page
is now twenty four, so that the page loads quicky with
default settings.

Pagination links now include the location and state so that
these filter parameters are not lost when changing pages.
The maximum page count is determined by running find_jobs
in "count" mode with the same query parameters as the main
query.