DB: do not allow - in column names
Used a single regular expression "valid_column" for checking column
names when preparing SQL statements dynamically.  Also test columns
separately to catch the case of a zero-length column name being
supplied.  Updated the unit test to check for private column names
being given to set_obs_info.  Created private versions of set_obs_info
and set_tilelist which take a database lock context.  These are then
called from add_job so that the info and tile list are set in the
same database context as that in which the job is added.  This is
necessary for the unit test to act as expected: otherwise jobs can
be half-inserted and then subsequent tests using the same tag will
fail due to the tag uniqueness constraint rather than the obs info
errors which are being tested.  Also added a test of the (edited)
set_obs_info and get_obs_info methods.