Start script to investigate unauthorized errors
A new module has been added to provide access to the OMP
and JCMT databases (in Sybase at JAC).  The new script has
been implemented as a sub-command of a general "jsa_proc"
tool.  It is expected to be too time consuming a process
to be done as part of the web interface.  It currently
checks the release dates to identify which jobs would
be expected to fail due to the presence of a future
release date in our database.