DB: fix find_jobs count when used with obsqueries
The find_jobs method, as written, cannot use joins when the count option
is active.  With "GROUP BY" the database returns multiple count values
rather than the single total we are looking for, but without "GROUP BY"
it will count jobs multiple times if there are multiple matches in the
joined table.

This commit addresses the problem by replacing the join to the obs
table with a where clause matching the job identifier to a subquery of
job identifiers matching the obs query in the obs table.  The method
therefore no longer needs to be given a join string.