CADC: alter TAP query to check only for CL 0
The unauthorized error investigation script was finding some
observation IDs to be in CAOM-2 when the raw data cannot be
found in Advanced Search.  This was because the observation
was present but with only reduced data planes.  This commit
updates the ADQL used in the TAP query used for observation
ID searches to restrict it to calibration level zero (raw
instrumental) data.