Directories: split job directories into 1000s
The job ID based directories are now broken into subdirectories
in groups of 1000 in order to prevent the creation of excessively
large directories.  Set the (minimum) job ID decimal length to
nine digits (three blocks of three) because using this system
for processing jobs other than the JSA HEALPix-based product
creation was suggested at yesterday's JCMT meeting.  There are
over a million jobs in the CADC data processing system (not
all of which are for JCMT) so the numbering scheme was set up to
accommodate job identifiers of this order of magnitude.