Web: add prev / next links to job info page
This works by storing the user's job query in their session,
and then passing it to the new job_prev_next method while
preparing the job info page.  The existing pagination system
is used, with the addition of two new parameters to the
render_page_control macro.  These allow the description
of the link targets to be changed ("job" instead of "page")
and the first / last links to be hidden.