Web: use "URL" query in session for prev / next
Storing the arguments to find_jobs in the session was troublesome
because they could include objects which the session system can
not serialize by default.  It was decided that storing the "URL"
query would be better than storing the find_jobs query arguments.
The code for preparing the job query from the query parameters
has been separated out into a new function "job_search".

Updated the parameter handling of the job list page to make the
URL query parameters consistent with the template context parameters.
The parameters are now at the same level (a "query" dictionary)
in the template context rather than being split between the top
level and an "obsquery" dictionary.  Made the spelling of "qa_state"
consistent everywhere (not "qastate").

The render_choices template macro does not need to be given the
parameters from the template context (since it can see it),
so removed them as macro arguments.