Web: allow automatic hyphenation of table headings
This is to allow some of the wider tables to be viewed on narrower
screens without scrolling sideways.  Firefox requires the language
to be specified, but appears to ignore the Content-Language header,
therefore it has also been specified via the HTML element's lang
attribute.  Chrome apparently does not support automatic
hyphenation.