Web: add SAMP broadcast buttons for FITS outputs
Currently assume that any FITS output file is an image and
show a button allowing the SAMP mtype image.load.fits to be
broadcast.  The broadcasting code is based on the sampjs
example file "broadcaster.js".  sampjs itself has not been
added to the repository as its license is unclear.