Web: select SAMP broadcast mtype by file name
Extracted the code which produces file lists for the web interface
because it was duplicated between the job_info and job_qa_info
scripts.  Added the SAMP mtype to the FileInfo namedtuple, so that
it can be put in a data-mtype attribute of the broadcast buttons.
The SAMP broadcast JavaScript then uses this mtype for the message
which it sends.