Web: quick hack to QA page for JSA catalog
On the job QA page, this shows the preview for the first parent job if
there are no preview images for a catalog. It also adds in the parent
.fits files to the QA page, as these are useful to send to Gaia and
examine when odd things are detected in the job. This will need a
better fix in the future, but is needed asap to do QA on legacy
catalogs.