Web: implement state_prev check on changing state
The prepare_change_state method supplies the state_prev argument
to db.change_state.  However the state_prev value given was
extracted from the database directly before making this change
as a placeholder for checking that the state did not change
since the user loaded the web page from which they are making
the state change.  This has now been implemented so that the
forms which post to this method should now include the previous
state, allowing the check to be performed.