Web: use error pages for state change failures
Using the new ErrorPage exception class allows more of the
logic for QA state changes to be moved out of the Flask
wrapper layer.  The error pages should also be more noticeable
than a flashed message, so failures should be hard to miss
when performing a series of QA assessments.