CADC: check disk space for e-transfer
Added a new entry to the [disk_limit] section of the configuration
file to configure the amount of free space required on the
e-transfer partition for more files to be transferred.  The limit
is currently set to 500GiB, i.e. about half of the normal available
space on that partition.