Web: redirect to error summary with same query
On changing job state from the error summary, now redirect back
to the error summary with the same query parameters.  This is
done by getting the full path requested including query parameters
(request.full_path) and passing it through to the error summary
template for use in the "url" hidden parameter.