Scripts: add command-based help system to jsa_proc
The jsa_proc script now turns off the automatic docopt help handling
and handles the --help option explicitly.  This allows it to provide
sub-command based help text.  The routine to do this uses the docstring
for the sub-command functions.  The existing text from the file
docstring has been moved to the corresponding functions, and the
docstrings for the other sub-command functions have been updated
for use with the new help system.