CADC: add ptransfer stamp file to aid clean-up
This file records the PID and start time of the ptransfer job which
created each "proc" directory.

Files are now organised within the "proc" directory by stream so that
they can be moved back to the appropriate directories on clean-up.
Sub-directories were used instead of logging the files to the stamp file,
or other log file, in case the process is terminated between moving
a file and the stamp/log entry being written to disk.