Action: always create output directory for job
Now instead of leaving it up to dpCapture to create the output directory
(in the case that there was previously a directory for this job), always
make the directory.  This means that dpCapture doesn't need to do this,
and that custom jobs do not need to worry about creating the transfer
directory.