web: Apply error filter to first error in log
So that we can apply the error filter to logs for jobs which
aren't actually in the ERROR state specifically, now search through
the log for the first ERROR state entry.  The error filter is then
applied to that log entry.  This means we can search by the actual
error of jobs marked as "won't work" (for example) rather than by
the comment entered when that state was set.  The previous state
constraint is now applied in the filter rather than the find_errors_logs
when finding jobs in states other than ERROR.