Submit: more directly compare parent job lists
Compare new and old parent job lists including filters rather than
comparing the jobs and filters separately.  This is so that we
would detect changes of which filter is applied to which job.

Also since the "parents" list is now only ever used as a set, convert
it to a set immediately on construction.