Web: sort output files for job information pages
Sort the output files before iterating through them to prepare
the file lists to be shown on the job information pages.  This is
mostly to ensure that the previews of various sizes are in the same
order such that clicking a 256px preview to see the larger version
actually shows the preview of the same product rather than one of
the other products.  (This still relies on no previews being missing
however.)

This should also make the output file list slightly easier to
read in general.